requirements.txt
llama-index
langchain
transformers
torch
streamlit
wikipedia-api
pdfplumber
sentence-transformers
faiss-cpu
